Skip iterations of disterbance regimes that are not requested in .csv files (issue #35)
If the requested .in file does not exist, that iteration is skipped for all climate conditions.

+++ StepWat_R_Wrapper_Parallel/StepWat.Wrapper.Code_V3.R
@@ -66,6 +66,10 @@ s<-site[1]
               for(intensity in graz_intensity ){
                 setwd(paste0(dist.directory))
                 dist.graz.name<-paste0("rgroup.freq",dst,".graz.",grz,".",intensity,".in")
+
+                if(!file.exists(dist.graz.name)){
+                  next
+                }
                 system(paste0("cp ",dist.graz.name," ",directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/"))
                 
                 setwd(paste0(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/"))
@@ -215,7 +219,12 @@ s<-site[1]
                 for (grz in graz.freq) {
                   for(intensity in graz_intensity ){
                     setwd(paste0(dist.directory))
-                    dist.graz.name<-paste0("rgroup.freq",dst,".graz",grz,".",intensity,".in")
+                    dist.graz.name<-paste0("rgroup.freq",dst,".graz.",grz,".",intensity,".in")
+
+                    if(!file.exists(dist.graz.name)){
+                      next
+                    }
+
                     system(paste0("cp ",dist.graz.name," ",directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/"))
                     
                     setwd(paste0(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/"))
